requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
autopage==0.5.2
backports.cached-property==1.0.2
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
build==0.10.0
CacheControl @ file:///home/conda/feedstock_root/build_artifacts/cachecontrol-split_1686152218986/work
cachetools==5.5.1
cachy @ file:///home/conda/feedstock_root/build_artifacts/cachy_1664983268779/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1666183775483/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1728479282467/work
-e git+https://github.com/MolecularAI/Chemformer.git@99bab0c0cc5200d61f63078204384bdff47b5733#egg=chemformer
cleo==2.1.0
cliff==3.10.1
clikit @ file:///home/conda/feedstock_root/build_artifacts/clikit_1676917709145/work
cmaes==0.11.1
cmd2==2.4.3
colorlog==6.9.0
crashtest==0.4.1
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1666563371538/work
cycler==0.11.0
dataclasses==0.6
deepchem==2.7.1
deepspeed==0.14.0
distlib @ file:///home/conda/feedstock_root/build_artifacts/distlib_1728557174656/work
dulwich==0.21.7
einops==0.6.1
exceptiongroup==1.2.2
fastapi==0.98.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1726613473834/work
flatbuffers==25.2.10
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.62.3
h5py==3.8.0
hjson==3.1.0
html5lib @ file:///home/conda/feedstock_root/build_artifacts/html5lib_1592930327044/work
huggingface-hub==0.16.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.2.3
jax==0.3.25
jeepney @ file:///home/conda/feedstock_root/build_artifacts/jeepney_1649085214306/work
joblib==1.3.2
jsonschema==4.17.3
keras==2.11.0
keyring==23.13.1
kiwisolver==1.4.5
libclang==18.1.1
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
more-itertools==9.1.0
msgpack @ file:///home/conda/feedstock_root/build_artifacts/msgpack-python_1654260645308/work
multidict==6.0.5
ninja==1.11.1.3
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
nvidia-ml-py==12.570.86
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
optuna==2.10.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandas==1.3.5
pastel @ file:///home/conda/feedstock_root/build_artifacts/pastel_1640899049124/work
pbr==6.1.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
Pillow==9.5.0
pkginfo @ file:///home/conda/feedstock_root/build_artifacts/pkginfo_1709561090460/work
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prettytable==3.7.0
protobuf==3.19.6
psutil==6.1.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.9.0
pydantic_core==2.14.6
pylev @ file:///home/conda/feedstock_root/build_artifacts/pylev_1641226376343/work
pynvml==11.5.3
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1724616129934/work
pyperclip==1.9.0
pyproject_hooks==1.2.0
pyrsistent==0.19.3
PySMILESutils @ git+https://github.com/MolecularAI/pysmilesutils.git@98b1f438578b624af3033e33435a8762c02d8093
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
PyTDC==0.4.1
pytest==7.4.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.2.3
pytz==2024.2
PyYAML==6.0.1
rapidfuzz==3.4.0
rdkit==2022.9.3
rdkit-pypi==2022.9.5
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1716354486713/work
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
SecretStorage @ file:///home/conda/feedstock_root/build_artifacts/secretstorage_1660605887060/work
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1698144360966/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.27.0
stevedore==3.5.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit @ file:///home/conda/feedstock_root/build_artifacts/tomlkit_1715185399719/work
torch==1.8.1
torchmetrics==0.11.4
tqdm==4.67.1
trove-classifiers==2025.1.15.22
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
urllib3==1.26.20
virtualenv==20.26.6
wcwidth==0.2.13
webencodings @ file:///home/conda/feedstock_root/build_artifacts/webencodings_1694681268211/work
Werkzeug==2.2.3
wrapt==1.16.0
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
